[DF-185] Fixed win32 project of SceneViewer
Former-commit-id: 487c2b93e2692e4e57f7138078df9d6f6a0499e6

+++ Projects/SceneViewer/Classes/ViewSceneScreen.cpp
@@ -187,7 +187,7 @@ void ViewSceneScreen::UpdateInfo(float32 timeElapsed)
     
     if(framesTime > INFO_UPDATE_TIME)
     {
-        int32 fps = frameCounter / framesTime;
+        int32 fps = (int32)(frameCounter / framesTime);
         info->SetText(Format(L"FPS: %d  Draw: %ldns  Update: %ldns", fps, drawTime / frameCounter, updateTime / frameCounter));
         
         framesTime -= INFO_UPDATE_TIME;
